TST: rearranged unit tests
